requirements.txt
django
psycopg2
djangorestframework
django-angular
django-registration
django-bootstrap3
django-formtools